requirements.txt
Django==1.10.5
Pillow==4.0.0
argparse==1.2.1
wsgiref==0.1.2
uWSGI==2.0.12
MySQL-python==1.2.5
pytz==2013.8
